requirements.txt
pygame
opencv-python
numpy
Pillow
ultralytics
gtts
pandas